backend/pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "hometrack"
version = "0.1.0"
description = ""
authors = ["David Hoover <hooverd@arizona.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-ninja = "^1.1.0"
boto3 = "^1.34.103"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
